% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{chart_satellites}
\alias{chart_satellites}
\title{Create a SNAPverse satellites flowchart diagram}
\usage{
chart_satellites(file_name, context = "all", selected = NULL,
  file_type = NULL, title = NULL, width = 350, height = NULL)
}
\arguments{
\item{file_name}{output file name, defaults to \code{"sv_td_all.svg"} for a topdown chart of the full verse.}

\item{context}{character, all statellites, utils packages only, or dev packages. See details.}

\item{selected}{character, vector of any selected packages intended to be highlighted vs. others. See details.}

\item{file_type}{the output file type.}

\item{title}{character, flowchart title.}

\item{width}{numeric.}

\item{height}{numeric.}
}
\value{
side effect of saving image file to disk.
}
\description{
Save a SNAPverse satellite packages flowchart diagram to disk.
The default \code{context = "all"} creates a flowchart of all satellite packages.
Other options include \code{"utils"} and \code{"all"}, which create flowcharts for
the three utils (app, map, and snap) packages and the SNAPverse development packages (meta and site), respectively.
If \code{selected} is not \code{NULL}, selected packages are given a fixed color and any unselected are gray.
If \code{NULL}, each context group is colored uniquely.
Other arguments are passed to \code{DiagrammeR::export_graph}.
}
\examples{
\dontrun{
chart_satellites(file_name = "sv_satellites_all.svg",
  context = "all", title = "SNAPverse satellite packages")
chart_satellites(file_name = "sv_satellites_utils.svg",
  context = "utils", title = "SNAPverse utils satellites")
chart_satellites(file_name = "sv_satellites_utils_app.svg",
  context = "utils", selected = "apputils",
  title = "SNAPverse utils satellites")
chart_satellites(file_name = "sv_satellites_dev.svg",
  context = "dev", title = "SNAPverse development satellites")
}
}
