% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{chart_verse}
\alias{chart_verse}
\title{Create a SNAPverse flowchart diagram}
\usage{
chart_verse(file_name, context = "complete", title = NULL,
  file_type = NULL, width = 350, height = NULL)
}
\arguments{
\item{file_name}{output file name.}

\item{context}{character, type of verse flowchart diagram. Defaults to \code{"complete"}.}

\item{title}{character, flowchart title.}

\item{file_type}{the output file type.}

\item{width}{numeric.}

\item{height}{numeric.}
}
\value{
side effect of saving image file to disk.
}
\description{
Save a SNAPverse flowchart diagram to disk. The default \code{context = "complete"} is the only option at this time.
Other arguments are passed to \code{DiagrammeR::export_graph}.
}
\examples{
\dontrun{
x <- "Simplified overview of the SNAPverse"
chart_verse(file_name = "sv_all.svg", title = x)
chart_verse(file_name = "sv_all.png", title = x, width = 800)
}
}
