% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{make_pkg_icons}
\alias{make_pkg_icons}
\title{Make package icons}
\usage{
make_pkg_icons(base_path = ".", width = 400, height = 200)
}
\arguments{
\item{base_path}{output directory.}

\item{width}{numeric, icon width.}

\item{height}{numeric, icon height.}
}
\value{
side effect of writing files.
}
\description{
Make package icon svg thumbail files.
}
\examples{
\dontrun{make_pkg_icons()}
}
