% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapsite.R
\name{verse_diagram}
\alias{verse_diagram}
\title{Create a SNAPverse diagram}
\usage{
verse_diagram(verse = "complete", file_name = "sv_td_all.svg",
  file_type = NULL, title = NULL, width = 350, height = NULL)
}
\arguments{
\item{verse}{character, type of verse flowchart diagram. Defaults to \code{"complete"}.}

\item{file_name}{output file name, defaults to \code{"sv_td_all.svg"} for a topdown chart of the full verse.}

\item{file_type}{the output file type.}

\item{title}{character, flowchart title.}

\item{width}{numeric.}

\item{height}{numeric.}
}
\value{
side effect of saving image file to disk.
}
\description{
Save a SNAPverse diagram to disk. The default \code{verse} is \code{"complete"}. It is the only option at this time. More coming soon.
Other arguments are passed to \code{DiagrammeR::export_graph}.
}
\examples{
\dontrun{verse_diagram(title="Simplified overview of the SNAPverse")}
}
